+++ app/scheduler.py
@@ -62,4 +62,5 @@ def run_loop():
 
             time.sleep(config["CHECK_INTERVAL"])
         else:
+            controller.turn_off_plug()
             sleep_until_next_start(config["START_TIME"])